query changes

+++ src/database/queries/sessionAttendees.js
@@ -112,12 +112,32 @@ exports.findAll = async (filter, tenantCode, options = {}) => {
 
 exports.unEnrollAllAttendeesOfSessions = async (sessionIds, tenantCode) => {
 	try {
+		const recordsToDelete = await SessionAttendee.findAll({
+			where: {
+				session_id: { [Op.in]: sessionIds },
+				tenant_code: tenantCode,
+			},
+			raw: true, // optional, returns plain objects
+		})
+
+		// If nothing found
+		if (!recordsToDelete.length) {
+			return {
+				deletedCount: 0,
+				deletedRecords: [],
+			}
+		}
+
 		const destroyedCount = await SessionAttendee.destroy({
 			where: {
 				session_id: { [Op.in]: sessionIds },
 				tenant_code: tenantCode,
 			},
 		})
+		return {
+			deletedCount: destroyedCount,
+			deletedRecords: recordsToDelete,
+		}
 
 		return destroyedCount
 	} catch (error) {
